requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.52.0
httpx==0.27.2
numpy==2.0.2
faiss-cpu==1.13.0


